Add files via upload
The entire study and article written on my analyses of wing deformation in beetle wings as a function of their vein placement and cross-sectional shape. This was achieved by creating a rotation matrix about the leading edge to infer trailing edge displacement and by documenting lift force generated as per each wing.